AuthViews: restructured code
