+++ src/views/ActivateView.js
@@ -6,7 +6,7 @@ import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import Typography from '@material-ui/core/Typography';
 
-import PasswordField from './PasswordField.js';
+import PasswordField from '../components/PasswordField.js';
 
 
 const styles = theme => ({
